Address PR review: document tlib adapters and tighten CPU ownership.
- map_host_region Safety: tlib TCG dereferences the host pointer
- CODE_WINDOW / bind_memory / map_memory: TODO for phase-D MappedRegion wiring
- TlibSeat: panic on a second live acquire instead of waiting
- Breakpoint mirrors: comment why kernel ids and tlib address sets both exist
- run_quantum: take u32 (tlib_execute); kernel clamps the time budget
- Reset-after-execute skip is cfg(test) only (UNIT_TEST_RESET_TLIB_ON_NEXT_NEW)
- read_reg / sync_breakpoints: document CLI/GDB inspect path
- build.rs: x86 / x86_64 hosts only

+++ sim-kernel/src/cpu.rs
@@ -151,12 +151,13 @@ pub trait Cpu: Send {
 
     /// Execute up to `max_instructions` guest instructions.
     ///
-    /// Mirrors `tlib_execute(max_insns)`. On return, map the exit code with
+    /// `u32` matches `tlib_execute(max_insns)`. The kernel clamps the virtual-time
+    /// budget to `u32::MAX` before calling this. On return, map the exit code with
     /// [`map_tlib_exit`]: breakpoint → `EXCP_DEBUG` → [`StopReason::Breakpoint`].
     /// Implementations must either retire a non-zero number of instructions or
     /// return a stop reason. A zero-instruction quantum with no stop is treated
     /// as [`StopReason::Halt`].
-    fn run_quantum(&mut self, max_instructions: u64) -> Quantum;
+    fn run_quantum(&mut self, max_instructions: u32) -> Quantum;
 
     fn read_reg(&self, id: RegId) -> Result<u64, SimError>;
     fn write_reg(&mut self, id: RegId, value: u64) -> Result<(), SimError>;
